requirements.txt
exif==1.6.0
ffmpeg_python==0.2.0
Flask==3.0.3
Flask_Cors==4.0.1
geopy==2.4.1
moviepy==1.0.3
numpy==1.26.4
openai==1.42.0
opencv_python==4.10.0.84
Pillow==10.4.0
pillow_heif==0.16.0
protobuf==5.27.3
pydub==0.25.1
PyExifTool==0.5.6
pyrootutils==1.0.4
scikit_learn==1.5.1
torch==2.2.2
tqdm==4.66.4
transformers==4.41.1
google-cloud-storage
google-cloud-vision
grpcio==1.64.1
Flask==3.0.3
pyyaml